Clean up the look of the filter bar and extract it into its own view, also fix progress bar style not showing
